This Update Server Side Pagination Fixed. Search is back to working.
There was problem with the library. i was using a library which was intended for arrays parameters., but my datatables was not sending parameters in arrays.. after using different library problem fixed.

This is the Library i am using now.
https://github.com/cryogenix/Ignited-Datatables/blob/master/application/libraries/Datatables.php

Next Update Update and Delete Buttons will work. Right now Edit and Delete is not working ATM.