1. TabID is being sent 0, No Matter what Tab i choose. 2.Now this Second One is not the Programmatic Error but logical you might say. As if i choose the Tab Users then if want to select the Parent Menu then Only Menus it should show me from the Users Tab as i have Selected the Users Tab. But it shows all the Menus from all the Tabs.. Which Gets Confusing.
Right Now These are the Two Problems i should Concentrate On.

Yes Still there are Two Problems Remaining, I just updated the MenuName Field which i first forgot to mention about that in problems above as form do get added but to what menu? Menu Name must be assigned to the menu to which the form will be Added.